+++ full-moon/src/ast/mod.rs
@@ -329,6 +329,11 @@ pub enum Value {
     #[cfg(feature = "roblox")]
     #[display(fmt = "{}", "_0")]
     IfExpression(IfExpression),
+    /// An interpolated string, such as `` `hello {"world"}` ``
+    /// Only available when the "roblox" feature flag is enabled.
+    #[cfg(feature = "roblox")]
+    #[display(fmt = "{}", "_0")]
+    InterpolatedString(InterpolatedString),
     /// A table constructor, such as `{ 1, 2, 3 }`
     #[display(fmt = "{}", "_0")]
     TableConstructor(TableConstructor),
@@ -2306,7 +2311,7 @@ impl Ast {
     /// The entire code of the function
     ///
     /// ```rust
-    /// # fn main() -> Result<(), Box<std::error::Error>> {
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
     /// assert_eq!(full_moon::parse("local x = 1; local y = 2")?.nodes().stmts().count(), 2);
     /// # Ok(())
     /// # }