+++ full-moon/src/lib.rs
@@ -24,6 +24,9 @@ mod private;
 mod short_string;
 mod util;
 
+#[cfg(feature = "roblox")]
+mod tokenizer_luau;
+
 pub use short_string::ShortString;
 
 use std::fmt;